Change parseFloate to pareseFloat JS file and * to × in HTML file
Made an error in using the parseFloat method by putting an extra
'e' at the end. This made calculator screen to clear any number
once I try to use any operator or equal to symbol.

This took me almost 24hrs to find bug because I forgot to use the
developer console to check. This was a good lesson by the way.